+++ packages/terra-form-radio/src/RadioField.jsx
@@ -6,6 +6,10 @@ import classNamesBind from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
 import uniqueid from 'lodash.uniqueid';
 import VisualyHiddenText from 'terra-visually-hidden-text';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import {
+  VALUE_UP, VALUE_DOWN, VALUE_RIGHT, VALUE_LEFT,
+} from 'keycode-js';
 import styles from './RadioField.module.scss';
 
 const cx = classNamesBind.bind(styles);
@@ -109,13 +113,18 @@ const RadioField = (props) => {
     legendAttrs.className,
   ]);
 
+  const fieldSetId = `terra-radio-group-${uniqueid()}`;
   const legendAriaDescriptionId = `terra-radio-field-description-${uniqueid()}`;
   const helpAriaDescriptionId = help ? `terra-radio-field-description-help-${uniqueid()}` : '';
   const errorAriaDescriptionId = error ? `terra-radio-field-description-error-${uniqueid()}` : '';
   const ariaDescriptionIds = `${legendAriaDescriptionId} ${errorAriaDescriptionId} ${helpAriaDescriptionId}`;
 
+  const isSafari = navigator.userAgent.indexOf('Safari') !== -1 && navigator.userAgent.indexOf('Chrome') === -1;
+  const isEdge = navigator.userAgent.indexOf('Edg') !== -1;
+  const LegendGroup = (isSafari || isEdge) ? 'div' : 'legend';
+
   const legendGroup = (
-    <legend id={legendAriaDescriptionId} className={cx(['legend-group', { 'legend-group-hidden': isLegendHidden }])}>
+    <LegendGroup id={legendAriaDescriptionId} className={cx(['legend-group', { 'legend-group-hidden': isLegendHidden }])}>
       <div {...legendAttrs} className={legendClassNames}>
         {isInvalid && <span className={cx('error-icon')} />}
         {required && (isInvalid || !hideRequired) && (
@@ -132,13 +141,38 @@ const RadioField = (props) => {
           )}
         {!isInvalid && <span className={cx('error-icon-hidden')} />}
       </div>
-    </legend>
+    </LegendGroup>
   );
+
+  const handleKeyDown = (event) => {
+    const radioGroup = document.getElementById(fieldSetId);
+    const radioItems = radioGroup.querySelectorAll('[type=radio]');
+    const itemIndex = Array.from(radioItems).indexOf(event.currentTarget);
+    if (event.key === VALUE_DOWN || event.key === VALUE_RIGHT) {
+      if (itemIndex === radioItems.length - 1) {
+        radioItems[0].focus();
+        radioItems[0].checked = true;
+      } else {
+        radioItems[itemIndex + 1].focus();
+        radioItems[itemIndex + 1].checked = true;
+      }
+    } else if (event.key === VALUE_UP || event.key === VALUE_LEFT) {
+      if (itemIndex === 0) {
+        radioItems[radioItems.length - 1].focus();
+        radioItems[radioItems.length - 1].checked = true;
+      } else {
+        radioItems[itemIndex - 1].focus();
+        radioItems[itemIndex - 1].checked = true;
+      }
+    }
+  };
 
   const content = React.Children.map(children, (child) => {
     if (child && child.type.isRadio) {
       return React.cloneElement(child, {
-        inputAttrs: { ...child.props.inputAttrs, 'aria-describedby': ariaDescriptionIds },
+        inputAttrs: {
+          ...child.props.inputAttrs, 'aria-describedby': ariaDescriptionIds, onKeyDown: handleKeyDown,
+        },
       });
     }
 
@@ -146,7 +180,7 @@ const RadioField = (props) => {
   });
 
   return (
-    <fieldset {...customProps} required={required} className={radioFieldClasses}>
+    <fieldset id={fieldSetId} {...customProps} required={required} className={radioFieldClasses}>
       {legendGroup}
       {content}
       {isInvalid && error && <div id={errorAriaDescriptionId} className={cx('error-text')}>{error}</div>}